made set work technically and renamed placeholder in Add to nullpointer cuz thats what it is. made i in hash a regular int to prevent mismatch
on another note if my tests dont stop having an unresolved external i'm going to do nothing about it because choices made in anger cannot be undone

but by god i will be mad. i will be so fucking mad about it. positively fuming god damn. the garden. that's a game i have on my switch there is a capybara in it. maybe two it;s been a while

ok insane rambling over buh-bye !  Σ:3